+++ server/src/main/java/ru/practicum/shareit/booking/mapper/BookingMapper.java
@@ -2,8 +2,8 @@ package ru.practicum.shareit.booking.mapper;
 
 import org.springframework.stereotype.Component;
 import ru.practicum.shareit.booking.dto.BookingDto;
-import ru.practicum.shareit.booking.model.Booking;
 import ru.practicum.shareit.booking.dto.BookingDtoTime;
+import ru.practicum.shareit.booking.model.Booking;
 import ru.practicum.shareit.booking.model.Status;
 import ru.practicum.shareit.user.model.User;
 import ru.practicum.shareit.item.model.Item;
