+++ server/src/main/java/ru/practicum/shareit/booking/service/BookingService.java
@@ -1,8 +1,8 @@
 package ru.practicum.shareit.booking.service;
 
 import ru.practicum.shareit.booking.dto.BookingDto;
-import ru.practicum.shareit.booking.dto.BookingDtoTime;
 import ru.practicum.shareit.booking.model.Status;
+import ru.practicum.shareit.booking.dto.BookingDtoTime;
 
 import java.util.Collection;
