+++ server/src/main/java/ru/practicum/shareit/booking/service/BookingServiceImpl.java
@@ -5,7 +5,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import ru.practicum.shareit.booking.dto.BookingDto;
-import ru.practicum.shareit.booking.dto.BookingDtoTime;
 import ru.practicum.shareit.booking.mapper.BookingMapper;
 import ru.practicum.shareit.booking.model.Booking;
 import ru.practicum.shareit.booking.model.Status;
@@ -21,6 +20,7 @@ import ru.practicum.shareit.user.repository.UserRepository;
 import ru.practicum.shareit.item.model.Item;
 import java.time.LocalDateTime;
 import ru.practicum.shareit.booking.model.Booking;
+import ru.practicum.shareit.booking.dto.BookingDtoTime;
 import java.util.Collection;
 import java.util.List;
 import java.util.Objects;
